requirements/api.txt
# Django REST Framework
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-nested-routers==0.93.4

# CORS handling
django-cors-headers==4.3.1

# Filtering
django-filter==23.5

# API Documentation
drf-spectacular==0.27.0

# Rate limiting
django-ratelimit==4.1.0

# Admin theme
django-admin-interface==0.31.0
django-colorfield==0.11.0

# Caching and async
celery==5.3.4
redis==5.0.1
